requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
annotated-types==0.7.0
    # via pydantic
appdirs==1.4.4
    # via pint
arrow==1.3.0
    # via stravalib
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
flexcache==0.3
    # via pint
flexparser==0.3.1
    # via pint
idna==3.10
    # via requests
numpy==2.1.2
    # via pandas
pandas==2.2.3
    # via -r requirements/requirements.in
pint==0.24.3
    # via stravalib
pydantic==2.9.2
    # via stravalib
pydantic-core==2.23.4
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   pandas
python-dotenv==1.0.1
    # via -r requirements/requirements.in
pytz==2024.2
    # via
    #   pandas
    #   stravalib
requests==2.32.3
    # via stravalib
six==1.16.0
    # via python-dateutil
stravalib==2.0
    # via -r requirements/requirements.in
types-python-dateutil==2.9.0.20241003
    # via arrow
typing-extensions==4.12.2
    # via
    #   flexcache
    #   flexparser
    #   pint
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
